requirements.txt
face-recognition==1.3.0
face-recognition-models==0.3.0
Flask==2.0.3
Flask-Cors==3.0.10
Flask-Login==0.6.1
h5py==3.6.0
imutils==0.5.4
Jinja2==3.0.3
keras==2.8.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
keras-vggface @ git+https://github.com/rcmalli/keras-vggface.git@719ad75625fe51436b9a9479114b3d136847e11a
numpy==1.22.3
opencv-contrib-python==4.1.2.30
opencv-python==4.5.4.60
Pillow==9.0.0
pyTelegramBotAPI==4.5.0
tensorflow==2.8.0
Werkzeug==2.0.3
